requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
diff-match-patch==20230430
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
django-import-export==4.1.1
django-rest-swagger==2.2.0
django-unfold==0.33.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
importlib_resources==6.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
pypstk==0.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
setuptools==70.2.0
simplejson==3.19.2
sqlparse==0.5.0
swagger-spec-validator==3.0.4
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
